refactor firewall step to accomodate new postures step
