+++ web/messages/en/location.json
@@ -15,6 +15,8 @@
   "add_location_step_access_control_description": "Assign user groups with access permissions to this location.",
   "add_location_step_firewall_label": "Firewall",
   "add_location_step_firewall_description": "The default policy defines how to handle traffic not covered by ACL rules.",
+  "add_location_step_posture_check_label": "Posture check",
+  "add_location_step_posture_check_description": "Define who can connect to your network.",
   "add_location_firewall_no_license_subtitle": "Firewall is available only in Business and Enterprise plans. You can still create location now and activate Firewall later by upgrading your plan.",
   "add_location_service_prelogon_title": "Pre-logon connection.",
   "add_location_service_prelogon_content": "The VPN connects at system boot and disconnects after user login, useful for one-time authentication with external identity providers like Active Directory.",
@@ -49,8 +51,8 @@
   "add_location_mfa_internal_content": "Uses the MFA methods configured in your Defguard profile.",
   "add_location_mfa_external_title": "External Identity Provider Authentication",
   "add_location_mfa_external_content": "Requires configuring an external identity provider in the settings, such as Google, Microsoft Entra ID, Okta, or JumpCloud.",
-  "add_location_step_posture_check_label": "Posture check",
-  "add_location_step_posture_check_description": "Define who can connect to your network.",
+  "add_location_postures_assign": "Assign posture check to newly created location",
+  "add_location_postures_dont_assign": "Don’t add any posture check for location",
   "location_add": "Add Location",
   "location_col_gateway_status": "Gateway status",
   "location_col_gateway_ip": "Gateway IP",
